% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itlt_models.R
\name{bslme}
\alias{bslme}
\title{Calculate test statistic using mixed effects model (LRT based model)}
\usage{
bslme(D.traj, D.cov, tf, nknot = 5, p.value = 0.05, degree = 3)
}
\arguments{
\item{D.traj}{dataset of purely longitudinal observations in wide format}

\item{D.cov}{covariate matrix of intercept, treat, grp, treat * grp}

\item{tf}{time frame}

\item{nknot}{knot number}

\item{p.value}{significance value required for a node to be splitted, default
is 1}

\item{degree}{degree of spline coefficients, default is 3}
}
\value{
(LRT) chi^2 test statistic of df(nknot) under the mixed effects model
}
\description{
Mixed effects based model tests trajectories by testing model fitting.
Trajectories are assumed to follow mixed effects models. Under the null
hypothesis there are only one treatment effect curve for both nodes. Under
the alternative, separate treatment effect curves are assumed for both nodes
}
