% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itlt_chi2normalize.R
\name{Chi2Nomalize}
\alias{Chi2Nomalize}
\title{Nomalizing a \chi^2 square statistic from df(n) to df(1)}
\usage{
Chi2Nomalize(t, n)
}
\arguments{
\item{t}{original test statistic}

\item{n}{original degree of freedom}
}
\value{
\chi^2 test statistic of df(1) with the same significance level
}
\description{
The pruning procedure is proposed under the assumption that test statistics at
each node are \chi^2(1) distributed.
}
