% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itlt_simulation.R
\name{ItltSimulation1}
\alias{ItltSimulation1}
\title{Simulation for longitudinal datasets}
\usage{
ItltSimulation1(N, tf, e = 0.2, p = 0.2, asigma = 0.5,
  seed = "1234", type = 1)
}
\arguments{
\item{N}{sample size}

\item{tf}{time frame}

\item{e}{error standard deviation}

\item{p}{missing value probability}

\item{asigma}{sd of subject specific random effect}

\item{seed}{seed for random number}

\item{type}{simulation setting (1: non-linearity, 2: periodicity,
3: non-linearity, 4: no heteorogeneity in treatment effect)}
}
\value{
a simulated longitudinal dataset
}
\description{
Simulation for longitudinal datasets
}
\examples{
sample <- ItltSimulation1(250, 1:12)
}
