% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itlt_tree.R
\name{ItltTree}
\alias{ItltTree}
\title{Implement tree structured subgroup identification on longitudinal datasets}
\usage{
ItltTree(data, fm, tf, split.covs, nknot = 5, p.value = 1,
  delta = 0.1, nCutpoints = 20, minsplit = 20, size.limit = 10,
  maxdepth = 3, details = "TRUE", alpha = 2.5, model = "bslme")
}
\arguments{
\item{data}{dataset in wide format, please see fm and time frame for column
name restrictions. Also data must have a column named "id", otherwise the first
column will be specified to be "id"}

\item{fm}{formula of the form "response name ~ treatment name". data must
contain column names of the response with all its time frames. For instance,
if fm = "y~treatment", tf = 1:12, then columns names in the dataset must
include y1, y2,..., y12. And these longitudinal observations will be used
as reponse for subgroup identfication}

\item{tf}{time frame}

\item{split.covs}{splitting covariates, vector of strings}

\item{nknot}{knot numbers, for multiple test based and mmp based methods,
nknot <= 0 means automatic selection of knot numbers using GCV}

\item{p.value}{significance level for a node to be considered, default is 1}

\item{delta}{lower bound of the ratio of the size of the child node and the
parent node, default is 0.1}

\item{nCutpoints}{number of cutoff points for a continuous splitting covariate,
default is 20}

\item{minsplit}{minimum number of samples for a node to be considered for
splitting, default is 30}

\item{size.limit}{minimum size for a node, default is delta * minsplit}

\item{maxdepth}{maximum depth of the tree, default is 4}

\item{details}{whether details in the treee building process will be printed,
default is FALSE}

\item{alpha}{tuning parametering in the selection of knot numbers using
smoothing splines, specific for multiple test based method and MMP based method,
default is 2}

\item{model}{method to be implemented.}
}
\value{
identified tree of partynode type
}
\description{
Methods implemented include: mixed effects based method (LRT based method),
GEE based method, multiple test based method, multivariate multiple regression
(mmp) based method. mixed effects and GEE are for regular longitudinal datasets and
multiple test based and mmp based method are for intensive longitudinal datasets.
This function is based on partynode package. Also interaction trees by scalar
based methods (avg, last-k) are also implemented.But the scalar based method
needs to have the same data format (see data, fm, tf for details), otherwise
regular code for interaction tree could be used
}
\examples{
library(itlt)
data <- ItltSimulation1(250, 1:12, type = 2)
ItltTree(data, as.formula(y~treatment), 1:12, paste0("X", 1:10))
}
