% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itlt_knot_select.R
\name{knot.select}
\alias{knot.select}
\title{Selection of number of knots in multiple test base and mmp based methods}
\usage{
knot.select(D.traj, tf, alpha = 2)
}
\arguments{
\item{D.traj}{dataset consists purely of longitudinal observations}

\item{tf}{time frame}

\item{alpha}{penalization term in GCV}
}
\value{
mean and standard deviation of GCV of each possible knot number
}
\description{
For intensive longitudinal dataset (multiple test based method and mmp based
method), a preprocessing step using smoothing splines is used.Generalized
cross validation (GCV) is used for knot number selection.
}
