% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itlt_spline_preprocess.R
\name{spline.preprocess}
\alias{spline.preprocess}
\title{Use smoothing spline to preprocess original longitudinal data}
\usage{
spline.preprocess(D.traj, tf, nknot, alpha)
}
\arguments{
\item{D.traj}{dataset consists purely of original longitudinal observations}

\item{tf}{time frame}

\item{nknot}{knot number}

\item{alpha}{penalization term in GCV}
}
\value{
new longitudinal dataset with fitted spline coefficients
}
\description{
For multiple test based and MMP based methods, preprocess using smoothing
splines are the first step
}
